Move tar tests to tests/tar.
This exposes some interesting test data dependencies, where rpm and deb
depend on things that only pkg_tar should care about. I intend to fix
those in followup CLs instead of clutting this one. Those tests should
be improved to be more useful for their context, and might require a
more careful review than this pure refactoring.

Followups:
- make legacy rpm test use appropriate input files
- split pkg_tar out of pkg.bzl
