+++ pkg/legacy/tests/rpm/BUILD
@@ -17,7 +17,17 @@ load("//pkg:rpm.bzl", "pkg_rpm")
 
 pkg_rpm(
     name = "test-rpm",
-    data = ["//tests:archive_testdata"],
+    # TODO(aiuto): Replace the use of these .ar files with something more
+    # usefule to testing rpm builds.
+    data = [
+        "//tests:testdata/a.ar",
+        "//tests:testdata/a_ab.ar",
+        "//tests:testdata/a_b.ar",
+        "//tests:testdata/a_b_ab.ar",
+        "//tests:testdata/ab.ar",
+        "//tests:testdata/b.ar",
+        "//tests:testdata/config",
+    ],
     debug = 1,
     release = "0",
     spec_file = "test_rpm.spec",
