+++ tests/tar/compressor.py
